havok_types_derive

Attribute Macro impl_flags_methods

#[impl_flags_methods]

Automatically implement [core::str::FromStr], [core::fmt::Display] in string notation according to Havok’s hkFlags<Enum, SizeType> XML notation.

In addition, [core::default::Default] is implemented.

Playground

Note

This attribute macro is automatically implemented by parsing bitflags! macro, so it cannot be implemented in a normal structure.